+++ lib/sdl3.dart
@@ -4,4 +4,5 @@ export 'sdl3/sdl_dart.dart';
 export 'sdl3/sdl_image.dart';
 export 'sdl3/sdl_mixer.dart';
 export 'sdl3/sdl_net.dart';
+export 'sdl3/sdl_shadercross.dart';
 export 'sdl3/sdl_ttf.dart';